change entitymanager and virtual thread executor config
